Allow any text-like input (e.g. email) to be styled as a text input.
So input types such as `email` and `tel` may be styled as text fields,
style any `input` which is the child of `.o-forms-input--text` as a
text input.

This means we don't have to provide individual selectors for a large
and ever changing set of input types.

See isse: https://github.com/Financial-Times/o-forms/issues/272