Allow the label of box radio buttons to be customised.
The classes `o-forms-input--saving` and `o-forms-input--saved`
add an icon and "Saving" or "Saved" copy to an element
`<span class="o-forms-input__state"></span>`.

The modifier classes and state element may be set using the
`setState` JS method, but may also be added manually.

As the `o-forms-input--saving` and `o-forms-input--saved` classes
add copy using pseudo elements the label of the state is not
customisable. E.g. "Saved" can't be changed to "Sent" where it
makes sense for the ui, and also the copy can't be localised.

Instead of include the copy for `saving` and `saved` states via
CSS, this PR proposes adding the copy via HTML. This may be
customised and localised. To do that whilst remaining backward
compatibile two new classes have been added `o-forms-input--loading`
(instead of `o-forms-input--saving`) and `o-forms-input--complete`
(instead of `o-forms-input--saved`). These add the saving/saved
icon without adding a pseudo element for the text label. Instead
the label is added with an extra span e.g.

```diff
<div class="o-forms-field">
	...
-	<span class="o-forms-input o-forms-input--radio-box">
+	<span class="o-forms-input o-forms-input--radio-box o-forms-input--loading">
		<div class="o-forms-input--radio-box__container">
			<label>
				<input type="radio" name="negative" value="Yes">
				<span class="o-forms-input__label">Yes</span>
			</label>
			<label>
				<input type="radio" name="negative" value="No"checked>
				<span class="o-forms-input__label o-forms-input__label--negative">No</span>
			</label>
		</div>
+		<span class="o-forms-input__state">
+			<span class="o-forms-input__state-label">Saving</span>
+		</span>
	</span>
</div>
```

We recommend users call `setState` rather than construct state
markup themselves. It has been updated to use this new markup.
Its third argument `iconOnly` is now an options map which includes
a setting for `iconOnly` and `iconLabel`, so the label may be
customsied.