Set input labels on date input demos.
Also remove `aria-hidden` from the "legend" of "field" elements.
This didn't break anything for screenreader users thanks to
the "aria-labelledby" attribute, but there doesn't seem to be a
good reason for it. Users have accidentally applied it to the
title of single inputs which have hidden content from screen
readers, so better to remove it if we can.