Improve guidance to try to make it harder to copy inaccessible markup.
- Move accessibility guidence within `README.md`.
- Expand `README.md` examples to include required ARIA labels.
- Simplify accessibility guidence by removing some of it (see below).

The guidence to hide the label of inputs in favour of `aria-label`,
when in a field of multiple inputs, has been removed. The guidence
was based on screen readers reading the label of the input twice
whilst tabbing through inputs but I was unable to replicate (using
VoiceOver on MacOS Catalina and Safari). _This should be confirmed
as part of code review._
https://github.com/Financial-Times/o-forms/blob/master/ACCESSIBILITY.md#aria-label--aria-hidden

Fixes: https://github.com/Financial-Times/o-forms/issues/364