Position state and error elements against inputs when labels are long. (#372)
Backports a fix from v8 to v7:
https://github.com/Financial-Times/o-forms/pull/306

When a label and input are inline, with label to the left and
input to the right, the state/error element is positioned on
the right under the input. If the label is long the status/error
message is pushed down away from the input due to absolute
positioning against their shared container.

Absolute positioning is used with a negative bottom value so the
state does not push content down the page when added dynamically.
The state sits within the fields margin instead. We can use
relative positioning and a negative margin instead. This keeps
the state close to the input when the label becomes long, but also
allows the state to avoid pushing other page content down by
negating the containers margin.