Add a white background to internal brand radio inputs.
Some users of internal tools take our radio buttons to be disabled
when placed on a `slate-white-15` background. This makes radio
buttons (round and box) always `white` for the internal brand.
It’ll continue to be transparent for the master brand, for use on
paper/white pages.

Closes: https://github.com/Financial-Times/o-forms/issues/389
Discussion: https://financialtimes.slack.com/archives/C01481FKWA2/p1614769119021000